// src/store/modules/global/types.ts
export enum TYPES {
  SET_VERSION_DATA = 'SET_VERSION_DATA',
  CLEAN_VERSION_DATA = 'CLEAN_VERSION_DATA',
  SET_SELECTVERSION_DATA = 'SET_SELECTVERSION_DATA',
  CLEAN_SELECTVERSION_DATA = 'CLEAN_SELECTVERSION_DATA',
  SET_LANGUAGES_DATA = 'SET_LANGUAGES_DATA',
  CLEAN_LANGUAGES_DATA = 'CLEAN_LANGUAGES_DATA',
  SET_SELECTLANG_DATA = 'SET_SELECTLANG_DATA',
  CLEAN_SELECTLANG_DATA = 'CLEAN_SELECTLANG_DATA'
}

export interface GLOBAL {
  versions: string[]
  selectVersion: string[]
  languages: string[]
  selectLang: string[]
}